requirements.txt
openai>=1.0
python-dotenv>=1.0
oracledb>=2.0
PyQt5>=5.15
PyYAML>=6.0
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
matplotlib>=3.7
